[Gradle] Working dir with @TempDir may not be set when all tests are skipped
However, @AfterAll will get executed, leading to unexpected
failure.

Checking for lateinit initialization should work.
If no workingDir were created cleanup should just early return.
There is no reason for it to fail.

This is fixup for KT-89398

^KT-89398 Verification Pending

+++ libraries/tools/kotlin-gradle-plugin-integration-tests/src/test/kotlin/org/jetbrains/kotlin/gradle/js/JsBrowserDebugSessionIT.kt
@@ -47,7 +47,9 @@ class JsBrowserDebugSessionIT : KGPBaseTest() {
     @AfterAll
     fun cleanWorkingDirIfPossible() {
         try {
-            workingDir.toFile().deleteRecursively()
+            if (this::workingDir.isInitialized) {
+                workingDir.toFile().deleteRecursively()
+            }
         } catch (e: IOException) {
             println("Failed to clean working directory: ${workingDir.toAbsolutePath()}, $e")
             // suppress
